+++ Changelog.md
@@ -1,7 +1,24 @@
+## [0.0.1] - UNRELEASED
+
+### Added
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+### Security
+
 ## [0.0.0] - 19/8/2024
 
 ### Added 
 
 - Commit inicial del proyecto (19/8/2024).
 - Script inicial del modelo (26/8/2024).
-- Tablero inical para el modelo (28/8/2024)
+- Tablero inical para el modelo (28/8/2024).
+- Escenario base 1 (28/8/2024).
+- Modelos Fuego, Punto de interes (29/8/2024).
+- Prefab Bombero, fuego y punto de interes (30/8/2024).